requirements.txt
streamlit
plotly
pandas
pandas-ta-classic
websocket-client